+++ src/test/java/test/pet/DeletePetTest.java
@@ -16,7 +16,7 @@ public class DeletePetTest {
 
     @Steps
     private PetEndpoint petEndpoint;
-    private long createdPetId;
+    private Integer createdPetId;
 
     @Before
     public void createPet() {
