+++ src/test/java/test/pet/UpdateExistingPetTest.java
@@ -17,7 +17,7 @@ public class UpdateExistingPetTest {
 
     @Steps
     private PetEndpoint petEndpoint;
-    private long createdPetId;
+    private Integer createdPetId;
 
     @Before
     public void createPet() {
