+++ src/test/java/test/pet/UpdatePetTest.java
@@ -17,7 +17,7 @@ public class UpdatePetTest {
 
     @Steps
     private PetEndpoint petEndpoint;
-    private long createdPetId;
+    private Integer createdPetId;
 
     @Before
     public void createPet() {
